Moved EntityType structs and enums from ztworldmgr to bfentitytype
